fixed env loading and tests folder structure
