+++ tests/conftest.py
@@ -1,17 +1,12 @@
 # pylint: disable=redefined-outer-name
 import os
-from pathlib import Path
 
 import pytest
-from dotenv import load_dotenv
 from fastapi import FastAPI
 from starlette.testclient import TestClient
 
 from service.api.app import create_app
 from service.settings import ServiceConfig, get_config
-
-BASE_DIR = Path(__file__).parent.parent
-load_dotenv(dotenv_path=BASE_DIR / ".env")
 
 
 @pytest.fixture
